requirements.txt
streamlit
google-generativeai
python-dotenv
sentence-transformers
pinecone